almost done

+++ include/structs.h
@@ -6,7 +6,7 @@
 /*   By: ahorling <ahorling@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2023/05/10 17:22:52 by ahorling      #+#    #+#                 */
-/*   Updated: 2023/05/15 22:17:35 by ahorling      ########   odam.nl         */
+/*   Updated: 2023/05/16 20:11:52 by ahorling      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,11 +17,10 @@
 # include <pthread.h>
 # include <unistd.h>
 
-
 /*This struct contains the basic information such as the dictated
 parameters, as well as the starttime of the process and
 if a philosopher has died.*/
-typedef struct	s_info
+typedef struct s_info
 {
 	size_t			num_of_philos;
 	size_t			time_to_die;
@@ -40,13 +39,13 @@ typedef struct	s_info
 to do their job, and is linked to the philosopher next in line.*/
 typedef struct s_philo
 {
-	int				number;
-	int				eat_count;
+	size_t			number;
+	size_t			eat_count;
 	bool			isdead;
 	size_t			last_ate;
 	pthread_t		thread;
 	pthread_mutex_t	*fork;
-	t_info	*info;
+	t_info			*info;
 	struct s_philo	*next;
 	struct s_philo	*prev;
 	bool			satiated;
